undo moves complete? i hope
update to how castling rights are stored. I added it under the Info struct inside of position.
I also changed it to be stored as a CastlingRights instead of an int. Likely an unnecessary change,
but it makes me happier with typing. This is why I had to update types.h